Rename slug to useSlug
