+++ src/app/landing/projects/ProjectItem.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import useSlug from '../../hooks/Slug'
+import useSlug from '../../hooks/useSlug'
 
 const ProjectItem = ({ project }) => {
   const slugify = useSlug()
